refactor: crypto utils
